src/app/modules/home/components/login-form/login-form.component.ts
import { Component } from '@angular/core';
import { AuthService } from 'src/app/core/services/auth.service';

@Component({
  selector: 'app-login-form',
  templateUrl: './login-form.component.html',
  styleUrls: ['./login-form.component.css'],
})
export class LoginFormComponent {
  formData = {
    username: '',
    password: '',
  };

  constructor(private authService: AuthService) {}

  onSubmit() {
    this.authService.login(this.formData).subscribe(() => {
      this.reloadPage();
    });
  }

  reloadPage() {
    window.location.reload();
  }
}
